Refactor tool_shed.util.tool_util into galaxy.
